move existing scripts to cosmos-scripts
