+++ core/engine/cosmos-scripts/start.js
